requirements.txt
# Core dependencies for Claude RAG system
setuptools>=65.0.0
wheel>=0.38.0

# API clients
anthropic>=0.8.0
openai>=1.0.0

# Local embedding models (no API key needed)
sentence-transformers>=2.2.0
torch>=2.0.0

# Data processing and ML
numpy>=1.24.0
scikit-learn>=1.3.0

# Utilities
python-dotenv>=1.0.0
tiktoken>=0.5.0

# Vector database integration
pinecone-client>=3.0.0 